+++ validate_email/smtp_check.py
@@ -3,58 +3,12 @@ from smtplib import (
     SMTP, SMTPNotSupportedError, SMTPResponseException, SMTPServerDisconnected)
 from typing import List, Optional, Tuple
 
-from dns.exception import Timeout
-from dns.rdatatype import MX as rdtype_mx
-from dns.rdtypes.ANY.MX import MX
-from dns.resolver import (
-    NXDOMAIN, YXDOMAIN, Answer, NoAnswer, NoNameservers, resolve)
-
-from .constants import HOST_REGEX
 from .email_address import EmailAddress
 from .exceptions import (
-    AddressNotDeliverableError, DNSConfigurationError, DNSTimeoutError,
-    DomainNotFoundError, NoMXError, NoNameserverError, NoValidMXError,
-    SMTPCommunicationError, SMTPMessage, SMTPTemporaryError)
+    AddressNotDeliverableError, SMTPCommunicationError, SMTPMessage,
+    SMTPTemporaryError)
 
 LOGGER = getLogger(name=__name__)
-
-
-def _get_mx_records(domain: str, timeout: int) -> list:
-    'Return the DNS response for checking, optionally raise exceptions.'
-    try:
-        return resolve(
-            qname=domain, rdtype=rdtype_mx, lifetime=timeout,
-            search=True)  # type: Answer
-    except NXDOMAIN:
-        raise DomainNotFoundError
-    except NoNameservers:
-        raise NoNameserverError
-    except Timeout:
-        raise DNSTimeoutError
-    except YXDOMAIN:
-        raise DNSConfigurationError
-    except NoAnswer:
-        raise NoMXError
-
-
-def _get_cleaned_mx_records(domain: str, timeout: int) -> list:
-    """
-    Return a list of hostnames in the MX record, raise an exception on
-    any issues.
-    """
-    records = _get_mx_records(domain=domain, timeout=timeout)
-    to_check = list()
-    host_set = set()
-    for record in records:  # type: MX
-        dns_str = record.exchange.to_text().rstrip('.')  # type: str
-        if dns_str in host_set:
-            continue
-        to_check.append(dns_str)
-        host_set.add(dns_str)
-    result = [x for x in to_check if HOST_REGEX.search(string=x)]
-    if not result:
-        raise NoValidMXError
-    return result
 
 
 class _SMTPChecker(SMTP):
@@ -209,7 +163,7 @@ class _SMTPChecker(SMTP):
     def check(self, hosts: List[str]) -> bool:
         """
         Run the check for all given SMTP servers. On positive result,
-        return `True`, else raise exceptions described in `mx_check`.
+        return `True`, else raise exceptions described in `smtp_check`.
         """
         for host in hosts:
             LOGGER.debug(msg=f'Trying {host} ...')
@@ -223,41 +177,27 @@ class _SMTPChecker(SMTP):
             raise SMTPTemporaryError(error_messages=self.__temporary_errors)
 
 
-def mx_check(
-        email_address: EmailAddress, debug: bool,
+def smtp_check(
+        email_address: EmailAddress, mx_records: list, debug: bool,
         from_address: Optional[EmailAddress] = None,
-        helo_host: Optional[str] = None, smtp_timeout: int = 10,
-        dns_timeout: int = 10, skip_smtp: bool = False) -> bool:
+        helo_host: Optional[str] = None, smtp_timeout: int = 10) -> bool:
     """
-    Returns `True` as soon as the any server accepts the recipient
-    address.
+    Returns `True` as soon as the any of the given server accepts the
+    recipient address.
 
     Raise an `AddressNotDeliverableError` if any server unambiguously
     and permanently refuses to accept the recipient address.
 
     Raise `SMTPTemporaryError` if the server answers with a temporary
-    error code when validity of the email address can not be
-    determined. Greylisting or server delivery issues can be a cause for
-    this.
+    error code when validity of the email address can not be determined.
+    Greylisting or server delivery issues can be a cause for this.
 
     Raise `SMTPCommunicationError` if the SMTP server(s) reply with an
     error message to any of the communication steps before the recipient
     address is checked, and the validity of the email address can not be
     determined either.
-
-    In case no responsible SMTP servers can be determined, a variety of
-    exceptions is raised depending on the exact issue, all derived from
-    `MXError`.
     """
-    from_address = from_address or email_address
-    if email_address.domain_literal_ip:
-        mx_records = [email_address.domain_literal_ip]
-    else:
-        mx_records = _get_cleaned_mx_records(
-            domain=email_address.domain, timeout=dns_timeout)
-    if skip_smtp:
-        return True
     smtp_checker = _SMTPChecker(
         local_hostname=helo_host, timeout=smtp_timeout, debug=debug,
-        sender=from_address, recip=email_address)
+        sender=from_address or email_address, recip=email_address)
     return smtp_checker.check(hosts=mx_records)
